Finish (hopefully) refactoring
